+++ frontend/src/App.jsx
@@ -9,7 +9,6 @@ import ProfileCard from "./pages/profile";
 import Dashboard from "./pages/dashboard";
 import PrivateRoute from "./utils/private-route";
 import MatchingForm from "./pages/matchingForm";
-import UserProfile from "./pages/userprofile";
 
 function App() {
   return (
